refactor: Phase 1 — extract types.ts and create barrel index.ts
- Move all shared type definitions (Query, Where, ParentScope, IncludeMap,
  PropertyMetadata, RelationMetadata, ModelMetadata, etc.) from Ad4mModel.ts
  into types.ts
- Ad4mModel.ts re-exports types via 'export * from ./types' for backward compat
- Update surreal-utils.ts and decorators.ts to import from ./types (breaks
  circular import type dependency)
- Create model/index.ts barrel re-exporting all public symbols
- Update core/src/index.ts to use single './model' barrel import

All 119 tests pass (115 passed, 4 skipped).
