+++ core/src/model/Ad4mModel.ts
@@ -6,6 +6,14 @@ import { makeRandomId, PropertyOptions, Model, getPropertiesMetadata, getRelatio
 import { singularToPlural, pluralToSingular, propertyNameToSetterName, relationToAdderName, relationToRemoverName, relationToSetterName } from "./util";
 import { escapeSurrealString } from "../utils";
 import { formatSurrealValue as _formatSurrealValue, compileWhereClause } from "./surreal-utils";
+import type {
+  WhereCondition, Where, Order, ParentScope, IncludeMap, Query,
+  RelationSubQuery, GetOptions, AllInstancesResult, ResultsWithTotalCount,
+  PaginationResult, PropertyMetadata, RelationMetadata, ModelMetadata, ValueTuple,
+} from "./types";
+
+// Re-export all shared types so existing consumers of './Ad4mModel' are unaffected
+export * from "./types";
 
 // JSON Schema type definitions
 interface JSONSchemaProperty {
@@ -44,161 +52,6 @@ interface JSONSchemaToModelOptions {
   resolveLanguage?: string;
   local?: boolean;
   propertyOptions?: Record<string, Partial<PropertyOptions>>;
-}
-
-type ValueTuple = [name: string, value: any, resolve?: boolean];
-export type WhereOps = {
-  not: string | number | boolean | string[] | number[];
-  between: [number, number];
-  lt: number; // less than
-  lte: number; // less than or equal to
-  gt: number; // greater than
-  gte: number; // greater than or equal to
-  contains: string | number; // substring/element check
-};
-export type WhereCondition = string | number | boolean | string[] | number[] | { [K in keyof WhereOps]?: WhereOps[K] };
-export type Where = { [propertyName: string]: WhereCondition };
-type Order = { [propertyName: string]: "ASC" | "DESC" };
-
-/**
- * Discriminated union for parent-scoped queries.
- *
- * **Model form** (preferred) — predicate auto-resolved from the parent model's
- * relation metadata. Use `field` to disambiguate when the parent has multiple
- * relations targeting the same child class.
- *
- * **Raw form** — explicit predicate string, no metadata lookup.
- */
-export type ParentScope =
-  | { model: typeof Ad4mModel; id: string; field?: string }
-  | { id: string; predicate: string };
-
-/**
- * Describes which relations to eager-load when querying.
- *
- * Each value is either:
- * - `true` — hydrate the relation one level deep
- * - A `RelationSubQuery` — scoped sub-query (filter / sort / paginate / nested include)
- *
- * @example
- * ```typescript
- * // One level deep
- * { comments: true }
- *
- * // Sub-query: only the 5 most-recent comments
- * { comments: { order: { createdAt: 'DESC' }, limit: 5 } }
- *
- * // Nested eager-load
- * { comments: { include: { author: true } } }
- * ```
- */
-export interface IncludeMap {
-  [relation: string]: boolean | RelationSubQuery;
-}
-
-export type Query = {
-  /** Filter to instances that are the target of a link from a given parent. */
-  parent?: ParentScope;
-  properties?: string[];
-  include?: IncludeMap;
-  where?: Where;
-  order?: Order;
-  offset?: number;
-  limit?: number;
-  count?: boolean;
-};
-
-/**
- * Sub-query options for a specific relation inside an `IncludeMap`.
- *
- * Equivalent to `Query` without top-level scoping (`parent`) or `count`,
- * since the result set is already constrained to the linked relation.
- *
- * @example
- * ```typescript
- * await post.get({ include: { comments: { order: { createdAt: 'DESC' }, limit: 5 } } });
- * ```
- */
-export type RelationSubQuery = Omit<Query, 'parent' | 'count'>;
-
-/**
- * Options accepted by the instance `get()` method.
- *
- * A subset of `Query` — only hydration controls apply to a single known instance.
- */
-export type GetOptions = Pick<Query, 'include' | 'properties'>;
-
-export type AllInstancesResult = { AllInstances: Ad4mModel[]; TotalCount?: number; isInit?: boolean };
-export type ResultsWithTotalCount<T> = { results: T[]; totalCount?: number };
-export type PaginationResult<T> = { results: T[]; totalCount?: number; pageSize: number; pageNumber: number };
-
-/**
- * Metadata for a single property extracted from decorators.
- */
-export interface PropertyMetadata {
-  /** The property name */
-  name: string;
-  /** The predicate URI (through value) */
-  predicate: string;
-  /** Whether the property is required */
-  required: boolean;
-  /** Whether the property is read-only */
-  readOnly: boolean;
-  /** Initial value if specified */
-  initial?: string;
-  /** Language for resolution (e.g., "literal") */
-  resolveLanguage?: string;
-  /** Custom Prolog getter code */
-  prologGetter?: string;
-  /** Custom Prolog setter code */
-  prologSetter?: string;
-  /** Custom SurrealQL getter code */
-  getter?: string;
-  /** Whether stored locally only */
-  local?: boolean;
-  /** Transform function */
-  transform?: (value: any) => any;
-  /** Whether this is a flag property */
-  flag?: boolean;
-}
-
-/**
- * Metadata for a single relation extracted from decorators.
- */
-export interface RelationMetadata {
-  /** The relation name */
-  name: string;
-  /** The predicate URI (through value) */
-  predicate: string;
-  /** Custom SurrealQL getter code */
-  getter?: string;
-  /** Whether stored locally only */
-  local?: boolean;
-  /** Link direction: 'forward' for HasMany/HasOne, 'reverse' for BelongsToMany/BelongsToOne */
-  direction?: 'forward' | 'reverse';
-  /** Target model class thunk for hydration and type filtering */
-  target?: () => any;
-  /**
-   * Whether to auto-generate a conformance filter when `target` is set.
-   * Defaults to `true` — set to `false` to opt out of DB-level type filtering.
-   */
-  filter?: boolean;
-  /** Where clause for relation filtering (query DSL) */
-  where?: Where;
-}
-
-
-
-/**
- * Complete model metadata extracted from decorators.
- */
-export interface ModelMetadata {
-  /** The model class name from @Model */
-  className: string;
-  /** Map of property name to metadata */
-  properties: Record<string, PropertyMetadata>;
-  /** Map of relation name to metadata */
-  relations: Record<string, RelationMetadata>;
 }
 
 function capitalize(word: string): string {
